Работает, но некрасиво
