+++ js/main.js
@@ -1,4 +1,6 @@
+import './fetch.js';
 import './rendering-mini.js';
 import './rendering-full.js';
 import './upload.js';
 import './redactor.js';
+
